requirements.txt
astroid==2.4.2
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
coverage==5.3
decorator==4.4.2
dnspython==1.16.0
eventlet==0.28.0
Flask==1.1.2
Flask-Cors==3.0.9
Flask-SocketIO==4.3.1
Flask-SQLAlchemy==2.4.4
greenlet==0.4.17
idna==2.10
isort==5.5.2
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
monotonic==1.5
namegenerator==1.0.6
oauthlib==3.1.0
psycopg2-binary==2.8.6
pylint==2.6.0
PySocks==1.7.1
python-dotenv==0.14.0
python-engineio==3.13.2
python-socketio==4.6.0
requests==2.24.0
requests-oauthlib==1.3.0
rfc3987==1.3.8
six==1.15.0
SQLAlchemy==1.3.19
toml==0.10.1
tweepy==3.9.0
urllib3==1.25.10
validators==0.18.1
Werkzeug==1.0.1
wrapt==1.12.1
